add hyperopt option and reorganize repo
