+++ bofire/runners/api.py
@@ -1,0 +1,2 @@
+from bofire.runners.hyperoptimize import hyperoptimize
+from bofire.runners.run import run
