+++ pyproject.toml
@@ -29,3 +29,4 @@ max-complexity = 18
 "bofire/data_models/api.py" = ["F401"]
 "bofire/data_models/surrogates/api.py" = ["F401"]
 "bofire/benchmarks/api.py" = ["F401"]
+"bofire/runners/api.py" = ["F401"]
